+++ CHANGELOG.md
@@ -11,5 +11,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Issue #4: Export project
 - Issue #5: Import file
 - Issue #6: Confirm panel to reset project